Famliy Tree Application
This Family Tree Application allows users to create a new tree, save this new tree and load the tree to edit. Each tree will under a root person, and his parents, spouse, and children will be saved in this tree. If his children got a spouse, these children will be the root person of the subtree.